- Add subtitle prop to header and use in recipients header - Render send to in a credenza - Recipients card is more like a table as designed
